Refactor Naming Of Core, Tests & Main README.md
